fix(vglite) Move scissor wrappers in lv_vglite_utils.h
