+++ src/draw/nxp/vglite/lv_draw_vglite_blend.c
@@ -35,6 +35,7 @@
 
 #if LV_USE_GPU_NXP_VG_LITE
 #include "lv_vglite_buf.h"
+#include "lv_vglite_utils.h"
 
 /*********************
  *      DEFINES
@@ -95,18 +96,6 @@ static inline void lv_vglite_set_translation_matrix(const lv_area_t * dest_area)
  * @param[in] dsc Image descriptor
  */
 static inline void lv_vglite_set_transformation_matrix(const lv_area_t * dest_area, const lv_draw_img_dsc_t * dsc);
-
-/**
- * Enable scissor and set the clipping box.
- *
- * @param[in] clip_area Clip area with relative coordinates of destination buffer
- */
-static inline void lv_vglite_set_scissor(const lv_area_t * clip_area);
-
-/**
- * Disable scissor.
- */
-static inline void lv_vglite_disable_scissor(void);
 
 #if VG_LITE_BLIT_SPLIT_ENABLED
 /**
@@ -555,19 +544,6 @@ static inline void lv_vglite_set_transformation_matrix(const lv_area_t * dest_ar
         vg_lite_scale(scale, scale, &vgmatrix);
     }
     vg_lite_translate(0.0f - dsc->pivot.x, 0.0f - dsc->pivot.y, &vgmatrix);
-}
-
-static inline void lv_vglite_set_scissor(const lv_area_t * clip_area)
-{
-    vg_lite_enable_scissor();
-    vg_lite_set_scissor((int32_t)clip_area->x1, (int32_t)clip_area->y1,
-                        (int32_t)lv_area_get_width(clip_area),
-                        (int32_t)lv_area_get_height(clip_area));
-}
-
-static inline void lv_vglite_disable_scissor(void)
-{
-    vg_lite_disable_scissor();
 }
 
 #if VG_LITE_BLIT_SPLIT_ENABLED
